requirements.txt
# TensorFlow ve Keras (Çalışan versiyon!)
tensorflow-macos==2.16.1; sys_platform == 'darwin' and platform_machine == 'arm64'
tensorflow==2.16.1; sys_platform != 'darwin' or platform_machine != 'arm64'
tensorflow-metal==1.1.0; sys_platform == 'darwin' and platform_machine == 'arm64'
numpy==1.26.4

# ML Kütüphaneleri
scikit-learn==1.3.0
pandas==2.1.0

# Görselleştirme
matplotlib==3.7.2
seaborn==0.13.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Görüntü İşleme
pillow==10.1.0

# Ses İşleme
librosa==0.10.1

# Yardımcı
requests